Refactor ProtectedRoutes.tsx
Now protected routes is a function that return a RouteObject
