+++ src/router/OrganizationProtectedRoute.tsx
@@ -1,7 +1,7 @@
 import {Outlet} from 'react-router-dom'
 import {useAccountHealthTools} from '~components/Account/use-account-health-tools'
 import {useAuth} from '~components/Auth/useAuth'
-import {NoOrganizationsPage} from '~components/Organization/Dashboard/NoOrganizations'
+import {NoOrganizationsPage} from '~components/Organization/NoOrganizations'
 
 // This protected routes are supposed to be inside of a AccountProtectedRoute
 // So no auth/loading checks are performed here
