+++ src/router/ProtectedRoutes.tsx
@@ -1,12 +1,24 @@
 import AccountProtectedRoute from '~src/router/AccountProtectedRoute'
 import OrganizationProtectedRoute from '~src/router/OrganizationProtectedRoute'
+import { RouteObject } from 'react-router-dom'
+import { SuspenseLoader } from '~src/router/SuspenseLoader'
 
-const ProtectedRoutes = () => {
-  return (
-    <AccountProtectedRoute>
-      <OrganizationProtectedRoute />
-    </AccountProtectedRoute>
-  )
-}
+const ProtectedRoutes = ({ children }: { children: RouteObject[] }) => ({
+  element: (
+    <SuspenseLoader>
+      <AccountProtectedRoute />
+    </SuspenseLoader>
+  ),
+  children: [
+    {
+      element: (
+        <SuspenseLoader>
+          <OrganizationProtectedRoute />
+        </SuspenseLoader>
+      ),
+      children,
+    },
+  ],
+})
 
 export default ProtectedRoutes
