+++ src/router/routes/process-create.tsx
@@ -1,30 +1,26 @@
-import {lazy} from 'react'
+import { lazy } from 'react'
 // These aren't lazy loaded since they are main layouts and related components
 import LayoutProcessCreate from '~elements/LayoutProcessCreate'
 import ProtectedRoutes from '~src/router/ProtectedRoutes'
-import {Routes} from '.'
-import {SuspenseLoader} from '../SuspenseLoader'
+import { Routes } from '.'
+import { SuspenseLoader } from '../SuspenseLoader'
 
-// const ProtectedRoutes = lazy(() => import('../ProtectedRoutes'))
 const ProcessCreate = lazy(() => import('~elements/dashboard/processes/create'))
 
 const ProcessCreateElements = [
   {
-    element: (
-      <SuspenseLoader>
-        <ProtectedRoutes />
-      </SuspenseLoader>
-    ),
-    children: [
-      {
-        path: Routes.processes.create,
-        element: (
-          <SuspenseLoader>
-            <ProcessCreate />
-          </SuspenseLoader>
-        ),
-      },
-    ],
+    ...ProtectedRoutes({
+      children: [
+        {
+          path: Routes.processes.create,
+          element: (
+            <SuspenseLoader>
+              <ProcessCreate />
+            </SuspenseLoader>
+          ),
+        },
+      ],
+    }),
   },
 ]
